python/src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["you"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.18.4"
opencv-python = "^4.2.0"
pillow = "^7.1.2"
matplotlib = "^3.2.1"
pathlib = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
